refactor: protocol 의 공통 메소드는 extension 으로 구현 및 파일 분리 #108
