+++ ONETHING_iOS/ONETHING_iOS/Source/View/Common/GiveUpWarningPopupView.swift
@@ -79,5 +79,4 @@ final class GiveUpWarningPopupView: NNConfirmPopupView {
             $0.centerX.equalToSuperview()
         }
     }
-    
 }
