+++ ONETHING_iOS/ONETHING_iOS/Source/ViewModel/Home/HabitManagingViewModel.swift
@@ -5,25 +5,14 @@
 //  Created by kimdo2297 on 2022/03/04.
 //
 
-import Foundation
+import UIKit
 
 import RxRelay
 import RxSwift
 
-protocol GiveUpWarningPopupViewPresentable: AnyObject {
-    var currentDayText: String? { get }
-    var subTitleTextOfGiveupWarningPopupView: NSAttributedString? { get }
-}
-
-protocol FailPopupViewPresentable: AnyObject {
-    var titleTextOfFailPopupView: String? { get }
-    var progressCountTextOfFailPopupView: String? { get }
-    var reasonTextOfFailPopupView: String? { get }
-}
-
-final class HabitManagingViewModel: GiveUpWarningPopupViewPresentable, FailPopupViewPresentable {
+final class HabitManagingViewModel: GiveUpWarningPopupViewPresentable {
     let menuRelay = BehaviorRelay(value: HabitManagingViewModel.Menu.allCases)
-    var habitInProgressModel: HabitResponseModel?
+    private(set) var habitInProgressModel: HabitResponseModel?
     let loadingSubject = PublishSubject<Bool>()
     let completeSubject = PublishSubject<Void>()
     private let apiService: APIService
@@ -60,6 +49,10 @@ final class HabitManagingViewModel: GiveUpWarningPopupViewPresentable, FailPopup
         }).disposed(by: self.disposeBag)
     }
     
+    func update(habitInProgressModel: HabitResponseModel?) {
+        self.habitInProgressModel = habitInProgressModel
+    }
+    
     var titleTextOfStartAgainPopupView: NSAttributedString? {
         guard let pretendardFont = UIFont.createFont(
             type: .pretendard(weight: .semiBold),
@@ -77,65 +70,6 @@ final class HabitManagingViewModel: GiveUpWarningPopupViewPresentable, FailPopup
             range: (titleText as NSString).range(of: "기존의 기록은\n영구적으로 삭제돼요.")
         )
         
-        let paragraphStyle = NSMutableParagraphStyle().then {
-            $0.lineSpacing = 4.0
-            $0.alignment = .center
-        }
-        attributeText.addAttribute(
-            .paragraphStyle,
-            value: paragraphStyle,
-            range: NSRange(location: 0, length: attributeText.length)
-        )
-        return attributeText
-    }
-    
-    var currentDayText: String? {
-        guard let diffDays = self.diffDaysFromStartToCurrent else { return nil }
-        
-        return String(diffDays + 1)
-    }
-    
-    private var diffDaysFromStartToCurrent: Int? {
-        guard let habitInProgressModel = self.habitInProgressModel,
-              let startDate = habitInProgressModel.startDate.convertToDate(format: "yyyy-MM-dd")
-        else { return nil }
-        
-        let formatter = DateComponentsFormatter().then {
-            $0.allowedUnits = [.day]
-            $0.unitsStyle = .positional
-        }
-        
-        guard let diffDaysStr = formatter
-                .string(from: startDate, to: Date())?
-                .components(separatedBy: CharacterSet.decimalDigits.inverted)
-                .joined(), let diffDays = Int(diffDaysStr) else { return nil }
-        
-        return diffDays
-    }
-    
-    var titleTextOfFailPopupView: String? {
-        return "아쉽지만\n습관은 여기까지!"
-    }
-    
-    var progressCountTextOfFailPopupView: String? {
-        guard let currentDayText = self.currentDayText else {
-            return nil
-        }
-
-        return "진행: \(currentDayText)일차"
-    }
-    
-    var reasonTextOfFailPopupView: String? {
-        return "사유: 습관 그만하기"
-    }
-    
-    var subTitleTextOfGiveupWarningPopupView: NSAttributedString? {
-        guard let pretendardFont = UIFont.createFont(type: .pretendard(weight: .semiBold), size: 15)
-        else { return nil }
-        
-        let titleText = "열심히 달려온\n지금의 습관을\n정말로 그만하시겠어요?"
-        let attributeText = NSMutableAttributedString(string: titleText,
-                                                      attributes: [.font: pretendardFont, .foregroundColor: UIColor.black_100])
         let paragraphStyle = NSMutableParagraphStyle().then {
             $0.lineSpacing = 4.0
             $0.alignment = .center
@@ -183,5 +117,11 @@ final class HabitManagingViewModel: GiveUpWarningPopupViewPresentable, FailPopup
                 return "습관 그만하기"
             }
         }
+    }
+}
+
+extension HabitManagingViewModel: FailPopupViewPresentable {
+    var reasonTextOfFailPopupView: String? {
+        return "사유: 습관 그만하기"
     }
 }
